Change Node.print to print the preserved original input line
The previous approach so far was that the class Node would parse each 
line and somehow encapsulate attributes etc. this is still being done 
(though of course not fully implemented yet), but in addition to this, 
each line from the original input is stored unchanged in self.contents 
(if the line is the opening tag of a child node, a reference to that 
Node object is stored instead).
Now, when Node.print (or by extension, Tree.print) is called, the 
function will not attempt to reconstruct the line with original syntax 
and so on, but just print the original line.
The intention is that in the future, methods will be defined with which 
properties of a node can be modified, and the node will internally 
modify the line string as well.