Overhaul modules.paths
This new implementation (which is also not yet finished) has a new API:
the module now outputs either only absolute paths (-a), only relative
paths (-r) or both if no argument is provided. Additionally, it allows
the user to set the delimiter between the paths (-d, default = '\n') and
to optionally shell escape them (-e). an option --format has also been
added but not yet implemented.