Convert tabs to spaces and ensure sensible whitespace
All tabs have been converted to 4 spaces each, and blank lines have been 
modified to have the same indent as the lines around them. That means 
that for example in a function definition, like

```
def foo():
    process()
    
    print()
```

the line between `process` and `print` will consist of four spaces (in 
other words: every blank line has the same indentation "as its 
context").

Now before you complain because you don't like whitespace-only lines: 
there are different schools of though on this. I was thinking for a long 
time about what the most Pythonic way to handle this would be, but PEP8 
isn't clear on this, and googling gave very varied results as well. In 
the end it seems to be up to taste (which is fine by me, since it's not 
awfully important either).
I like the approach introduced with this commit, because each empty line 
still provides some context about what function or class or block or 
whatever it belongs to.